Update patch set 1
Patch Set 1: Code-Review-2

This is the wrong solution: the fix to the problem is a lot easier of this massive copy&paste.

See the Issue 12720 that documents the breakage, happened in the replication plugin with the commit 00ff9091494ccd81744bde643c981ebdead57894.

The fix needs to happen in the replication plugin.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-2