Fix issue with replication when replicationDelay is set to zero
Issue is cause by race condition. When replicationDelay is set to zero
task immediately is triggered immediately. Source.requestRunway(FetchOne)
is removing task from pending as a first operation when task is starting.
But  without delay task will be added to pending when it is already running
and will never be removed from pending which cause further replication to
add refs to task which is already finished instead of creation new one.

Bug: Issue 13099
Change-Id: If8ede575ceb0ae50e8192198c1321f4f7a3e46a6