Make sure replication task clean up is triggered
To trigger replication task cleanup `ReplicationState.allScheduled`
variable must be set to true.  This is done via
'ReplicationState.markAllFetchTaskScheduled`. Make sure that
'ReplicationState.markAllFetchTaskScheduled` is called before that is
scheduled. If not and task is quick enought it will finish before
markAllFetchTaskScheduled and notificatyFinished will not clean up task
properly

Feature: Issue 13100
Change-Id: I7431b2275ddd357bb04e0578200c68d4a392a69d