Execute refs fetch in batches
When #replication.useCGitClient is set to true. Pull replication plugin
uses command line client to execute fetch. Length of the command is
limited by the operating system. If number of refs to fetch is greater then
replication.refsBatchSize, refs are going to be split into separate fetch
operations.

Feature: Issue 11605
Change-Id: If5f192a95bf5f0945b503f66dcdc012ce71539eb