Parallelise REST API fetch calls
Allow configure number of threads for fetch calls. Fetch calls can be an
expensive operation because every fetch call is a blocking operation
and is waiting for git fetch to finish. Running fetch calls one by one
will multiply the time of a fetch by the number of remote destinations.
This have significant impacting one the overall performance. By allowing
parallel run of fetch calls we can reduce the overall time spent on
fetching and improve the performance.

Feature: Issue 13340
Change-Id: I5c961af2047365f330a6cad229d9870c533fc0ed