Fix FetchRefReplicatedEvent remote URI
When triggered by an applyObject execution, the triggered
FetchRefReplicatedEvent contained the `instanceLabel` that was used in
the request rather than the actual remote the ref was fetched from.

Lookup the remoteName and retrieve the actual URI from the source, so
that consumers of the FetchRefReplicatedEvent have visibility on the
actual source URL rather than the remote name in the configuration.

Bug: Issue 14865
Change-Id: I425b3deabd33d68ccdb4121d0772c04660292e31