Fix pull-replication after the removal of Log4J from Gerrit
Utilize Flogger's logging tag feature instead of the Log4J's MDC.
This is needed after the removal of Log4J by I1d01ab0a5d3ad.

The solution is aligned with the equivalent change on the
replication plugin I0d18f89312e31.

A small disadvantage is that the trace context is a part of the
log message i.e. we cannot really control the layout. What was earlier:

[2020-05-04 ...] [c0b3774c] Missing repository could not be created when replicating <repo-url>

becomes now:

[2020-05-04 ...] Missing repository could not be created when replicating <repo-url> [CONTEXT PLUGIN="replication" fetchOneId="c0b3774c" ]

Bug: Issue 15579
Change-Id: I07f2cab60c4695390bd7b4ce6db9e71053d6758b