Receive stream events from events-broker directly
Gerrit EventBus was not designed to manage the receiver
failures, making the execution of replication unreliable
without the implementation of custom-retry logic based on
replication tasks.

Receive the pull replication events directly from the events-broker
which is capable of controlling the acking of messages and their
redeliver in case of failures.

This requires the additional parameter replication.streamEventsTopic
which points to the topic name to use for receiving events.

Change-Id: Ibffed26eda4ca852e2599ffe664e8d39e35e9dd3