Receive stream events from events-broker directly
Gerrit EventBus does not support yet the ability to retry
the delivery of events in case of failures, making the execution
of replication unreliable in case of network failures.

Receive the pull replication events directly form the events-broker
which is capable of controlling the acking of messages and their
redeliver in case of failures.

This requires the additional parameter replication.streamEventsTopic
which points to the topic name to use for receiving events.

Change-Id: Ibffed26eda4ca852e2599ffe664e8d39e35e9dd3