Support ApplyObject of non-commit refs
Enable to apply non-commit refs directly through
the ApplyObject action, instead of requiring the execution
of a git fetch.

Rationale: the git fetch is not optimised for replicating
non-commits, because it relies on the reachability graph
of commit objects. The transmission of a ref-BLOB pair
is straightforward and implies the simple insertion of
a byte[] with an object type associated with a ref.

Change-Id: I90b74a60f9077bbbad29c38c5035ea12ee216c71