Consider any HTTP 2xx response code from REST-API as success
When executing remote REST-API for either ApplyObject or Fetch
any 2xx status code (200, 201, 202, 203, 204, ...) should be
considered as a completion and not reported as a failure or retried.

Previously, the 202 was missed and potentially also others could
be missed in the future. There is no reason for considering a 2xx
as an unsuccessful result.

Change-Id: Id2e55cfe441c73983001fc76156a47e5fe08d97c