Always fallback to fetch when ApplyObject REST-API fails
Previously the Git fetch was invoked only when the ApplyObject
failed because of missing parent commit. However, there are other
cases when the ApplyObject may fail and *must* be covered by
a fallback to Git fetch.

One example of these failures is a forced update of a branch
or a lock failure.

Falling back to a fetch allows to complete the replication anyway
by relying on the Git protocol to resolve the complete fetch.

Change-Id: I2a7b94d0088c65dbade3e3bdc0beae14d46e6410