Introduce the apply-objects REST-API for the whole '/meta' chain
When the parent commit of a '/meta' chain of NoteDb changes is missing
falling back to the Git fetch command would not help, because of the
inability of the Git fetch negotiation to find a common base.

Introduce a new apply-objects REST-API that is capable of applying
an entire chain of commits on the linear '/meta' ref chain so that the
Git fetch can be avoided altogether.

Bug: Issue 16052
Change-Id: I7cfbeeaba2fe6190b99fe78c7631bb884abe321f