Fix issue with ref deletion and global-refdb
Do not reuse `com.google.gerrit.server.restapi.project.DeleteRef` class
to delete ref because it operates with an high-level repository manager.

Using the high-level repository manager may be an issue in come cases
where it is wrapped, like in the multi-site setup (see [1]), with a global-refdb
and therefore may not be able to complete the REST-API successfully.

[1] https://gerrit.googlesource.com/plugins/multi-site/

Original-Author: maczech@gmail.com
Bug: Issue 16128
Change-Id: I2034a48fe04489e2827de7c44bdb0e67de3f4665