Do not use timer.getStartTime() when propating timers metrics
When the Fetch API client propagates the start time of the
E2E metrics, the native Gerrit metric timer value cannot be used
because it is not based on Epoch but on an internal JVM-dependent
timing.

Use instead the default System.currentTimeMillis() which is guaranteed
to be based on the Epoch.

Change-Id: I2ec7e5dc40dcb54f2cafc2be32e6c60fd93738e5