Remove authorisation from PullReplicationFilter
Assuming that the caller of the pull-replication APIs
is an identified user is not something that the filter
should do: Gerrit has already authentication defined
in its filters and we should rely on them.

Without the check of identified user, the pul-replication
must trust internal users to call its own APIs, therefore
add the condition of being an internal user in the ACL
evaluation.

Bug: Issue 16298
Change-Id: If6e7b407e26643098033fccae92b9a77c71e957d