Remove authorisation from ProjectInitializationAction
Assuming that the caller of the pull-replication APIs
is an identified user is not something that the ProjectInitializationAction servlet should do: Gerrit has already authentication defined in its filters and we should rely on them.

Without the check of identified user, the pull-replication
must trust internal users to call its own APIs, therefore
add the condition of being an internal user in the ACL
evaluation.

Bug: Issue 16327
Change-Id: Iee192dfb74b45ef4fbb18aad0eaf20e774d6bb5e