Decouple the GitRepositoryManager used from the LocalDiskRepositoryManager
The repository manager to use may be different from the LocalDiskRepositoryManager.
Use a generic named binding so that it would be possible to replace
the implementation by configuration.

In case there is no binding of the LocalDiskRepositoryManager,
make the binding explicitly in the PullReplicationModule so that
the whole chain of injections will still work as before.

Change-Id: I49257b1121016cf0a7d157a117420299bd28d437