Remove replicateNoteDbMetaRefs from remoteConfiguration
The replicateNoteDbMetaRefs() method was introduced only on the
master branch of the replication plugin. The pull-replication
stable-3.6 branch was initially created out of master and therefore
contained this unwanted method.

Remove the method and make sure that the re-merge to master becomes
a NOOP.

Also ignore the test shouldReturnBadRequestWhenDeletingAnInvalidProjectName
because it would only work on Gerrit master.

Change-Id: I89066ccfa14db38fe666f34cf0c2a39009c2c75f