Add Init phase in primary and replica.
Add push replication in the primary.

Also configure the initial ACLs of the All-Projects repo
for allowing the pull-replication internal user to access
all refs of All-Users repository. Without that ACL the
pull-replication would corrupt the repository because it
would only replicate the refs/meta/config but not the
groups and users.

Change-Id: Id27d1af7208250795c6a723624890e3555b42160