Allow PullReplicationInternalUser group to be nested inside groups
The PullReplicationInternalUser group is generated by the
PullReplicationGroupBackend, dynamically loaded as a plugin.

Gerrit loads plugins at the end of the startup phase and therefore
disallow the use of the custom group backends for certain configs
such as the upload.allowGroup setting.

Enable the ability to nest the PullReplicationInternalUser group
inside other Gerrit groups which can then be used in upload.allowGroup.

Example:
Internal group: Replication
- Included group: pullreplication:internal-user

[upload]
  allowGroup = Replication

Bug: Issue 40015330
Change-Id: I4faa6281fc0008756ab52e4aaa9ba5f3319eb9b4