Do not try to delete inexistent refs
Delete ref throws NPE when ref marked to be deleted doesn't exist
in the local repository. This can happen if repo is out of sync
or ref was deleted via different channel (for example via apply object).

Deleting a non-existent ref will just be a NOOP.
The apply-object API will still return a 200.

Bug: Issue 15873
Change-Id: Ic945a44268d7998f312f067ed009ae8af221a63f