Allow FetchOne to retry the failed ref only
FetchOne may contain more than one ref to be fetched: the failure
of one of them could block the whole set of fetches, causing an
unwanted replication delay.

Enable the possibility for a FetchOne to partially complete,
leaving a leftover of failed refs that will be logged for tracking
purposes.

This new behaviour has the immediate advantage of not delaying
the replication of refs that have no replication issues which would
be instead all cancelled because part of a replication task with
permanent failure.

TODO: Unit and integration tests are still missing.

Bug: Issue 16742
Change-Id: I093ce3a780313ada18323af7aaca391791b93441