Allow FetchOne to be partially successful and enable the retry of failed ref
FetchOne may contain more than one ref to be fetched: the failure of one of them could
block the whole set of fetches, causing an unwanted replication delay for no valid reason.

Enable the possibility for a FetchOne to partially complete, leaving a leftover of failed
refs that can be rescheduled and retried individually.

This new behaviour has the immediate advantage of not delaying the replication of refs
that have no replication issues and still assure that the failed replication events are still
going to be rescheduled for retry.

TODO: Unit and integration tests are still missing.

Bug: Issue 16742
Change-Id: I093ce3a780313ada18323af7aaca391791b93441