Allow creation of new projects
This change [1] introduced a regression when creating a
new project.

Project creation is skipped since the Source#shouldReplicate
check for projects existance in the cache, hence the method
should NOT be invoked for new projects.

Check the source configuration, but skipping the cache.

Note that the logic is not taking into consideration if a project is hidden.
A project could potentially be created on the remote and be visible until
the correct 'refs/meta/config' isn't replicated from the source.

The same is currently happening for push-replication, hence there is
no regression between the two solutions.

[1]: https://gerrit-review.googlesource.com/c/plugins/pull-replication/+/363004

Change-Id: Idd1bd88b8c46d31993896640fd9f9ae071ee034d