Set the config files after Gerrit init step
In 3.5 the events-kafka plugin introduced a `sendStreamEvents`
property which defaults to `false` [1].

The events-kafka init step should check if the value is
not set before defaulting it. This is not the case at the moment.

Moving the configuration setting after the init will fix the issue
while waiting for a fix in events-kafka [Issue 16815].

[1]: https://gerrit.googlesource.com/plugins/events-kafka/+/refs/heads/stable-3.5/src/main/java/com/googlesource/gerrit/plugins/kafka/InitConfig.java#60

Change-Id: I078e9ba8eb84a83b70c84375aa92ed273f82ef37