Add extra logging when replication tasks are merged
When two replication tasks are combined they were not previously
logged and therefore it was difficult to understand how they
were rescheduled and executed.

Add extra logging with the information associated with the tasks
being combined, so that the Gerrit admin can have a better
understanding and easily troubleshoot replication issues.

Change-Id: I464405525cf14f4992165d0a82724dac0a9c795a