Run apply-object before the ref-updated stream event
The pull-replication plugin's promise is to run apply-object
to replicate the refs in msec, so that the overall end-to-end
operation runs synchronously. However, the ref-updated was
executed before the apply-object, leaving the uncertainty to
the stream event listener of the ref update being run on all
remotes.

Use the same EventListener interface used in the stream-event
listeners for making sure that when the ref-updated is triggered
the data has been already replicated everywhere.

Also amend the test environment for the broker scenario for
including the apply-object use-case as well, where the events
broker is used as fallback by putting the events-kafka plugin
to the bottom of the plugin loading priority.

Change-Id: If96082f4338164e1674bae657ab4184577d825f2