Allow additional refs fallback to apply-objects
Apply-objects REST-API that is capable of applying an entire chain
of commits on the ref chain so that the Git fetch can be avoided
altogether. Without this change only refs/changes/**/meta fallback
to the apply-objects. Add `replication.applyObjectsRefs` parameter
to allow additional refs fallback to the apply-objects when
`MissingParentObject` exception is thrown by the receiver.

Change-Id: If2b12dee0c552defd97d3119eadd47f3700dc710