Use delete-project plugin's repository deletion logic to delete a git repo
The previous implementation was relying on the replication
plugin's LocalFS class, which was failing to delete
directories hosted on NFS filesystem, because of a lingering
file handle.

We are therefore swapping that implementation with the one
from the delete-project plugin, which can effectively handle
repos hosted on NFS.

Worth pointing out that this change introduces inter-plugin
dependency with the delete-project plugin.

Bug: Issue 16730
Change-Id: Iea80a01b84ca72aef5de21db3a05e07a626db079