Remove project from replication node's project cache on project deletion
Even if a project is deleted from the local filesystem via pull-
replication, it remains in the node's project cache. This is
problematic, as subsequent operations (eg to recreate the deleted repo)
fail, because the replica thinks the project exists locally.

This change is therefore removing the project from the node's cache,
whenever a project is successfully deleted from disk. To do this, it
reuses the CacheDeleteHandler from the delete-project plugin code.

Bug: Issue 16730
Change-Id: I62e7653a8d78f25c1de1f95a5f8df34e9c6a181e