Abstract the implementation of ApplyObject
Decouple the actual implementation of the ApplyObject API
from the way that it is used in different parts of the code.

ApplyObject can be used in different scenarios and should be
allowed to be tested and mocked regardless of any of them.

This is a preparation work for adding more external filters
to the ApplyObject behaviour for managing additional logic.

Change-Id: If0713d6717118fd04d735e3967b6d0c1aee2d25e