Ignore remote ref-updated stream events for replication
The ref-updated happened on foreign Gerrit servers, which means
that the instance-id of the event is different than the local one,
should not be considered as a ref-update to action.

The pull-replication has a specific separate listener to
react on remote events and the ReplicationQueue should only
be used for triggering replication actions due to local ref-updates
only.

Bug: Issue 16924
Change-Id: I8cd27f756becf4c4c990641d31639f30918f4da9