Remove change indexing on ref-replicated event
The multi-site plugin has always been responsible for
propagating the Gerrit indexes events across the sites.
However, since I5373bb8149, the pull-replication took the
direction of also managing part of the reindexing.

The propagation of indexing events is a lot more complex
that just triggering a single change index action and need
visibility of all the other machinery needed by a multi-site
setup.

This is the proper solution to the Issue 16934 which has
been resolved with a workaround in I8ebec4e and I0a28e302ee.

Bug: Issue 16934
Change-Id: Icd41a6f074088acf0d4c36982eb7524706e0fd9b