Dont read git submodule commits Don't read git submodule commits during replication
Currently, any attempt to read/parse a git submodule commit will
throw a `org.eclipse.jgit.errors.MissingObjectException`, as the git
submodule's commit hash cannot be found in the repo.

The git submodule's commit can be identified as it has the special
`160000` file mode.

The solution is to exclude the git submodule's commit from being
read/parsed.

I believe everything is replicated correctly in the end, but it will
be nice to get rid of this exception.

Bug: Issue 16897
Change-Id: Ib21b9e19c5282f94226d7a27c25c5fe0f94b7279