Explain how to configure plugin on primary/replica side
The existing documentation was mostly a copy/paste from the replication
plugin. However, the replication plugin configuration is conceptually
simpler because it always runs on one side of the replication process
only. The pull-replication plugin runs on both sides of the replication
and the current state of documentation didn't provide enough guidance
how to configure each side and which of the options are supported on
which side.

Change-Id: I7fe8cebda8e665ff8f122a1f70fbc707ab551fbf