Add event creation time to the apply object payload
Add eventCreatedOn field to the apply object REST
API json payload. This allows to send timestamp of the
event which triggered the apply object call. This
information is useful for the receiver side to understand
which events were already processed.

Adding new field to the apply object REST API is backward
compatible. Json payload generated without eventsCreatedOn
is deserialised correctly and eventCreatedOn field is set
to 0. Previous version of pull-replication ignores extra
fields in the json payload.

Bug: Issue 16885
Change-Id: I73c66cfc07285d3e1c2d0c4e190d56e21599a89d