Replicate first commit tree objects
Set postOrderTraversal to true in the TreeWalk object when reading
revision data for commit without parents. When post order traversal
is enabled then the walker will return a subtree after it has
returned the last entry within that subtree. This fix the issue with
"Missing Tree Object" exception for commits without parents.

Bug: Issue 40015563

Change-Id: I83a8e329917109f7d14c6006ea7609083634a772
Signed-off-by: jianghaozhe1 <jhzgg2017@gmail.com>