Pull-Replication Slave can not use cgit client to fetch from Master use git protocol without authentication
What steps will reproduce the problem?
  1. We have a Gerrit Master(v3.6.1+) and a Slave Server using Pull-Replication(stable-3.6+) to do replicaiton.
  2. Set "replication.useCGitClient=true" on the Master/Slave server in replication.config.
  3. For performance, we use git protocol without authentication(no username and password) on Slave to fetch from Master, such as: "git fetch git://ip/xiaomi.git  +refs/changes/38/2338/meta:refs/changes/38/2338/meta", and failed.

What is the expected output?
 Slave can use cgit client fetch successfully from Master without authentication.

This is because when not set username and password, the [protected URIish appendCredentials(URIish uri, CredentialsProvider credentialsProvider)] method in [src/main/java/com/googlesource/gerrit/plugins/replication/pull/fetch/CGitFetch.java] will enter the first if condition.

We have solved this problem by a more strick check username and password to avoid entrying the first if condition in xiaomi.com, and it works well.

Bug: Issue 16980

Change-Id: Ibf2569b2d3289a33386ff08e2aa7d30e546534ac
Signed-off-by: jianghaozhe <jhzgg2017@gmail.com>