New metric to track fetch task scheduling latency
Currently, the `ReplicationQueue` tracks the time it took to schedule a
fetch task to the target node using the e2e latency metric. This is
confusing, for 2 reasons: the same metric is used by the target to track
how long it actually took to fetch the refs, and also because
semantically this does not make sense, what we want to measure here is
how long the http request took to be accepted by the target node.

Introduce a new `replication_fetch_task_scheduling_latency metric` and
use that in the source.

Bug: Issue 16987
Change-Id: Ia7c6f5da3df08c9c9291694ba38da2ad183232e7