Apply-objects endpoint accepts a list of refs
This is the first step in a chain of changes. The end goal is for
pull-replication to support batching of references into a single call.

To this end, this change modifies the apply-objects endpoint so that it
expects a list of references. For now, we naively just select the first
ref, as we never expect more than one. In the future, once we handle a
new (batch) event type, we will update that logic.

Bug: Issue 16982
Change-Id: Ibc500d93c2aacc51620410be54ad8315dc3b42ee