Adapt to addition of replicateNoteDbMetaRefs in replication plugin
In I653f2e0a5b66dcded40c11ca23d2d22c63ccc396 the
`remote.<NAME>.replicateNoteDbMetaRefs` option was added to the
replication plugin, but it was still missing in pull-replication.
This caused the build of the plugin to fail.

To satisfy the compiler SourceConfiguration.replicateNoteDbMetaRefs()
is now overridden. It is however not used, since for pull-replication
the receiving side will always be a Gerrit, which can use meta-refs.
Thus, it is unlikely that this option has any relevance for this
plugin.

Change-Id: I30456615c5f9dd49c9967ae083f72b58a7ac9522