Schedule fetch as a single fetch task for all refs in a batch
Following the introduction of the batch-fetch api [1], the work on the
recipient side was outstanding. The payload of the batch fetch request
contains a list of all the refs in the batch, for which we want to
schedule a git fetch task. As per existing behaviour, there is no
guarantee that all (async) refs will end up in the same request. For
sync refs, there will be one request per pef. This is suboptimal.

This change schedules all the refs in the batch into a single fetch
task, for both sync and async refs.

[1] https://gerrit-review.googlesource.com/c/plugins/pull-replication/+/378496

Bug: Issue 303112557
Change-Id: Ifd2a28a6337f762d89d744934f44b7f2b47e6e6a