Do not alter replication state counters when ref is already scheduled
Looks like there's a bug in the code; we are incrementing the
`ReplicationState` counters even when a ref is already scheduled. This
can cause problems, because these counters are used to determine when
the fetch task is fully done.

What we really should do is only increase the counters whenever a fetch
task is created, or if we update the fetch tasks's state with new refs.
I've added a few unit tests around this logic, given nothing existed so
far.

Bug: Issue 290456749
Change-Id: I7ebc5a13b6d8459bea979002bf6b013c4abd89cb