Allow replication of hidden projects
When replicateHiddenProjects is set to true, the replication
task should be scheduled even if the project state does not
allow reads.

Also document that projects and project_list cache TTL needs to
be kept to a minimum, for allowing the results of the replication
to be effective on the system.

With regards to the REST-APIs (apply-object/fetch) document that
the pull replication user needs to be added as the OWNER of the
projects for being able to unhide through replication.

Bug: Issue 291876050
Change-Id: Iaa65b58b98fa05570a06b4c31ed204e864fda826