Use PullReplicationFilter in Gerrit primary setup
The use of the Gerrit REST-API framework for the pull-replication
APIs is problematic because it relies on the underlying project
status and ACLs.

When performing replication, we need to allow the pull replication
user to access the API regardless of the current status or ACLs of
the project, hence the standard REST-API are not suitable.

Change-Id: I0c0217985f67098f913d496a57f68b44932d951e