Use PullReplicationFilter in Gerrit primary setup
The use of the Gerrit REST-API framework for the pull-replication
APIs is problematic because it relies on the underlying project
status and ACLs.
When performing replication, we need to allow the pull replication
user to access the API regardless of the current status or ACLs of
the project, hence the standard REST-API are not suitable.

The PullReplicationFilter was already implemented for the Gerrit
replicas, therefore it is already suitable for use with the
primary as well, after being made independent from the project
resource parsing.

Remove in the documentation the note about being owner of the
project or administrator for replicating hidden projects, because
the adoption of the PullReplicationFilter would bypass the ACLs
altogether when resolving the project URL component.

Change-Id: I0c0217985f67098f913d496a57f68b44932d951e