Add integration test for new project creation
The integration test suite had the use-case of removing
a project but not the one for creating one.

Because the integration test is still half-broken, due
to the lack of the 2nd replication endpoint, rely on
the direct instantiation of a FetchApiClient which would
have been invoked if the remote end had notified the
creation of a project.

This half-integration test still verifies one of the two
ends of creating a project which would highlight potential
issues.

Change-Id: I39b6d7837b63856e8c555b7b9b392feee72838af