Honour project properties when initialising a new project
The creation of a new project may happen with extra properties when triggered
through the API.

The new properties were not honoured at project creation time and the further
fetch of refs/meta/config were not triggered by any project update action and
therefore did not trigger any eviction or reload.
As a result, the project stayed in memory with a stale cache, until the cache
would have eventually expired in 1h (or other maxAge configured).

This change propagates all the project's refs/meta/config to the other
endpoints, so that the project is created with the fully populated
config the first time and the correct cached values. The logic is fully reused
from the apply object command and relies on the physical objects transferred
as payload from the source to the destination replica endpoints.

Bug: Issue 296854545
Change-Id: Ic3139ce59667000b4e45571bbf720992f3b2c793