Introduce replication queue metrics
The pull-replication queue has a lot of intermediate
states for the events fired and the replication tasks
allocated in the various queues.

Expose the internal status of the queue with fine-grained
counters that allow to understand what is the overall
status on every source.

The plugin reload mode needs to be different now because
it won't be possible anymore to have two plugins (old and new)
loaded at the same time at any point in time, otherwise
the metric names would conflict.

Change-Id: I42db2dfe8b8de9afc202f650ae42d164f9754c51